Set omissible in template string Bug fix on last named arg
